STRIPE IMPROVEMENTS:
REFactor: Consolidate Stripe-related functionality into a single module, removing the deprecated stripe-customer and stripe-events files. Update imports across the application to reflect the new structure, enhancing maintainability and clarity in the codebase.
